+++ src/media/font/namespace.rs
@@ -3,7 +3,7 @@
 //! Defines [`Fonts`].
 //
 
-#[doc = crate::_tags!(font)]
+#[doc = crate::_tags!(font namespace)]
 /// Built-in and embedded font resources.
 #[doc = crate::_doc_meta!{ location("media/font"), }]
 #[derive(Debug)]