+++ src/media/visual/image/format/netpbm/namespace.rs
@@ -8,7 +8,7 @@ use crate::ImageError::{InsufficientBuffer, InvalidImageSize, InvalidPixel};
 use crate::{Extent2, ImageResult, unwrap};
 use PnmFormat::{P1, P2, P3, P4, P5, P6};
 
-#[doc = crate::_tags!(image namespace)]
+#[doc = crate::_tags!(image codec)]
 /// Encoding and decoding for <abbr title="Portable anymap format">PNM</abbr> images.
 #[doc = crate::_doc_meta!{location("media/visual/image")}]
 ///